[BOOKINGSG-6783][JZ] Renamed files
